Create system time provider
