+++ server/src/main/java/io/deephaven/server/runner/DeephavenApiServerModule.java
@@ -6,7 +6,6 @@ package io.deephaven.server.runner;
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ElementsIntoSet;
-import io.deephaven.base.clock.Clock;
 import io.deephaven.chunk.util.pools.MultiChunkPool;
 import io.deephaven.configuration.Configuration;
 import io.deephaven.engine.updategraph.UpdateGraphProcessor;
@@ -25,7 +24,7 @@ import io.deephaven.server.table.TableModule;
 import io.deephaven.server.table.inputtables.InputTableModule;
 import io.deephaven.server.uri.UriModule;
 import io.deephaven.server.util.Scheduler;
-import io.deephaven.time.TimeProviderClock;
+import io.deephaven.time.TimeProvider;
 import io.deephaven.util.process.ProcessEnvironment;
 import io.deephaven.util.thread.NamingThreadFactory;
 import io.grpc.BindableService;
@@ -121,8 +120,7 @@ public class DeephavenApiServerModule {
             }
         };
 
-        return new Scheduler.DelegatingImpl(serialExecutor, concurrentExecutor,
-                new TimeProviderClock(Clock.systemUTC()));
+        return new Scheduler.DelegatingImpl(serialExecutor, concurrentExecutor, TimeProvider.systemUTC());
     }
 
     private static void report(final String executorType, final Throwable error) {
