+++ server/test/src/main/java/io/deephaven/server/test/FlightMessageRoundTripTest.java
@@ -14,7 +14,6 @@ import io.deephaven.barrage.flatbuf.BarrageMessageWrapper;
 import io.deephaven.barrage.flatbuf.BarrageSnapshotOptions;
 import io.deephaven.barrage.flatbuf.BarrageSnapshotRequest;
 import io.deephaven.barrage.flatbuf.ColumnConversionMode;
-import io.deephaven.base.clock.Clock;
 import io.deephaven.base.verify.Assert;
 import io.deephaven.engine.context.ExecutionContext;
 import io.deephaven.engine.liveness.LivenessScopeStack;
@@ -37,7 +36,7 @@ import io.deephaven.server.session.SessionServiceGrpcImpl;
 import io.deephaven.server.session.SessionState;
 import io.deephaven.server.session.TicketResolver;
 import io.deephaven.server.util.Scheduler;
-import io.deephaven.time.TimeProviderClock;
+import io.deephaven.time.TimeProvider;
 import io.deephaven.util.SafeCloseable;
 import io.deephaven.auth.AuthContext;
 import io.grpc.ServerInterceptor;
@@ -108,8 +107,10 @@ public abstract class FlightMessageRoundTripTest {
 
         @Provides
         Scheduler provideScheduler() {
-            return new Scheduler.DelegatingImpl(Executors.newSingleThreadExecutor(),
-                    Executors.newScheduledThreadPool(1), new TimeProviderClock(Clock.systemUTC()));
+            return new Scheduler.DelegatingImpl(
+                    Executors.newSingleThreadExecutor(),
+                    Executors.newScheduledThreadPool(1),
+                    TimeProvider.systemUTC());
         }
 
         @Provides
